Support pure edm4hep output from ddsim

+++ JugTrack/src/components/ParticlesFromTrackFit.cpp
@@ -139,7 +139,7 @@ namespace Jug::Reco {
                                           covMomentum,
                                           static_cast<float>(parameter[Acts::eBoundTime]),
                                           timeError,
-                                          boundParam.charge()};
+                                          static_cast<float>(boundParam.charge())};
             track_pars->push_back(pars);
           }
